requirements.txt
APScheduler>=3.6,<4
flask>=1.0.2,<2
gunicorn>=19.8.1,<20
paho-mqtt>=1.3.1,<2
PyYAML>=4.2b1,<5
requests>=2.20,<3
RPi.GPIO>=0.6,<1; platform_machine == 'armv6l'
rpyc>=4.0.2,<5
smbus2>=0.2.0,<1
SQLAlchemy>=1.3,<2

-e .
